Cheddar 1.1 Release
A full release that supports oAuth, my convention-over-configuration
approach, and an intelligent library that utilises the latest and
greatest coding practices.

In essence, it's the best we can give Cheddar in PHP form. It is my
implementation on how an oAuth library should be, and it has
respectively reflected that.

It may be subject to debates, but I welcome all opinions and
contributions. Happy coding!